Modified search to query a single file.
