+++ example/run_minimize.sh
@@ -1,0 +1,14 @@
+echo "EXAMPLE: RUN MINIMIZE"
+
+echo -e "\n%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%\n"
+echo "===COMPILING==="
+echo -e "\n%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%\n"
+make clean -C
+make minimize -C
+
+echo -e "\n%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%\n"
+echo "===MINIMIZATION==="
+echo -e "\n%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%\n"
+../minimize \
+  -i "k35C_bef_mininimization.fa" \
+  -o "k32C_af_mininimization.fa"
